LPS-76193 Move into activator package, it's not used anywhere else
